+++ cascalog-core/src/clj/cascalog/logic/options.clj
@@ -25,7 +25,7 @@
    :spill-threshold nil
    :reducers nil
    :name ""
-   :stat-fn nil})
+   :stats-fn nil})
 
 (defn careful-merge
   "Semigroup that keeps the right value of it's not nil or not equal