+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 - in-memory planner!
 - `:name` option allows you to supply a name to your flow via (:name "some name")
 - Schema validation on internal functions
-- New stats interface via `:stat-fn`
+- New stats interface via `:stats-fn`
 - Options are now printed in `explain-query` readout
 
 ## 2.0.0